backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emoji-backend"
version = "0.1.0"
description = "Backend service for the emoji credentials platform"
authors = [{ name = "Factory Droid", email = "factory@example.com" }]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0,<0.112.0",
    "pydantic>=2.8.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "sqlmodel>=0.0.21,<0.0.22",
    "bcrypt>=4.1.2,<5.0.0",
    "python-jose[cryptography]>=3.3.0,<3.4.0",
    "uvicorn[standard]>=0.30.0,<0.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "httpx>=0.27.0,<0.28.0",
    "ruff>=0.5.0,<0.6.0"
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["app"]
